// archive/pre_v3_reset_2026_02_09/src/gen/config.ts
export type TerrainConfig = {
  elevationFrequency: number;
  elevationOctaves: number;
  elevationPersistence: number;
  elevationLacunarity: number;
  continentalFrequency: number;
  seaLevel: number;
  shoreBand: number;
  warpFrequency: number;
  warpAmplitude: number;
  moistureFrequency: number;
  moistureOctaves: number;
  moisturePersistence: number;
  moistureLacunarity: number;
  contourInterval: number;
  contourStrength: number;
  riverRegionSize: number;
  riverSeedsPerRegion: number;
  riverStartElevation: number;
  riverStepLength: number;
  riverMaxSteps: number;
};

export type VegetationConfig = {
  treeGridSize: number;
  treeRenderMargin: number;
  forestMinDensity: number;
  forestDenseThreshold: number;
  treeMinRadius: number;
  treeMaxRadius: number;
};

export type SettlementConfig = {
  cellSize: number;
  minVillageDistance: number;
  suitabilityThreshold: number;
  targetMoisture: number;
  maxCoastSearch: number;
  preferredCoastMin: number;
  preferredCoastMax: number;
};

export type RoadConfig = {
  regionSize: number;
  nearestNeighbors: number;
  maxConnectionDistance: number;
  loopChance: number;
  majorWidth: number;
  minorWidth: number;
  localWidth: number;
  routeStep: number;
  maxCurvatureOffset: number;
};

export type HouseConfig = {
  spacing: number;
  sideChance: number;
  minSetback: number;
  maxSetback: number;
  minWidth: number;
  maxWidth: number;
  minDepth: number;
  maxDepth: number;
  maxSlope: number;
};

export type ChunkConfig = {
  pixelSize: number;
  maxCachedChunks: number;
  sampleStep: number;
  featureMargin: number;
  superchunkSpanChunks: number;
  maxCachedSuperchunks: number;
  generationBudgetMs: number;
  maxChunkBuildsPerFrame: number;
  prefetchLookaheadChunks: number;
  prefetchLateralChunks: number;
  enableSeamValidation: boolean;
  seamColorTolerance: number;
};

export type DebugLayerConfig = {
  showContours: boolean;
  showRivers: boolean;
  showWaterMask: boolean;
  showMoisture: boolean;
  showForestMask: boolean;
  showRoads: boolean;
  showVillages: boolean;
  showParcels: boolean;
  showHouses: boolean;
};

export type WorldConfig = {
  seed: string;
  terrain: TerrainConfig;
  vegetation: VegetationConfig;
  settlement: SettlementConfig;
  roads: RoadConfig;
  houses: HouseConfig;
  chunk: ChunkConfig;
  debug: DebugLayerConfig;
};